Working on common codebase for poll, starting with legacy
